finished implementing most of the basic functionality

+++ src/graphql/resolvers/recommendation-resolver.js
@@ -10,7 +10,7 @@ export default {
             //check if the recommendation is a valid restaurant in the local db
             const found = await Restaurant.findById({_id: restaurant})
             if(found){
-                const rec = await Recommendation.create({body, author: user._id, restaurant});
+                const rec = await Recommendation.create({body, author: user._id, restaurant, location: {type: "Point", coordinates: [found.location.coordinates[0], found.location.coordinates[1]]}});
                 const me = await requireAuth(user);
                 let recs = me.recs;
                 await recs.push(rec)
@@ -27,31 +27,52 @@ export default {
         }
     },
     
-    getRecommendations : async (_, args, {user}) => {
-        await requireAuth(user);
+    getNearbyRecommendations: async (_, { coords, distance }, {user}) => {
+        if(coords){
+            let lat, lon;
         try {
-            return Recommendation.find({})
-        }catch (err){
-            throw err
+            //location should come in as an array of strings
+            lon = parseFloat(coords.split(",")[0]);
+            lat = parseFloat(coords.split(",")[1]);
+        } catch(err){
+            throw new Error("Invalid location format!")
+        }
+        
+        let rests = await Recommendation.find({
+          location: {
+           $nearSphere: {
+            $geometry: {
+             type: "Point",
+             coordinates: [lon, lat]
+            },
+            $maxDistance: distance
+           }
+          }
+         });
+
+        return rests;
+        }else {
+            throw new Error("Undefined coordinates!");
         }
     },
     
     deleteRecommendation: async (_, {_id}, {user}) => {
         await requireAuth(user);
         try {
-            return Recommendation.deleteOne({_id})
+            const rec = await Recommendation.findOne({_id})
+            if(rec.author._id == user._id){
+                return Recommendation.deleteOne({_id})
+            }else {
+                throw new Error("Unauthorized!");
+            }
         } catch (err){
             throw err
         }
     },
     
-    getRestaurants: async (_, {location}, user) => {
+    getRecommendations: async (_, args, {user}) => {
         await requireAuth(user);
-        
-    },
-    
-    getNearbyRecommendations: async (_, {location}, user) => {
-        await requireAuth(user);
-        
+        return Recommendation.find({})
     }
+
 }
